fixed iv booster not working
in previous versions, we relied on the IVs not existing, which isn't a great idea. Looks like 1.6+ causes the IVs to be created with an empty set by default. Probably fishing. I left the "if it doesn't exist, create it" as a backup, but now we review the existing list, look to see if there's that many perfects already, then shore up the difference if necessary. This can result in greater than the points in perfects still, as unset IVs get rolled before spawn.